Move lexer into its own module
